colab/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arduino-colab-kernel"
version = "0.1.2"
description = "Jupyter magics for Arduino/ESP32"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICEN[CS]E*"]
authors = [
    { name = "Jiri Konecny", email = "jiri.konecny@vsb.cz" }
]
keywords = ["arduino", "esp32", "jupyter", "magics", "notebook"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: Education",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "pyserial",
    "ipykernel",
    "ipython",
]

[project.urls]
Homepage = "https://github.com/sgtkingo/ArduinoColab"
Repository = "https://github.com/sgtkingo/ArduinoColab"
Issues = "https://github.com/sgtkingo/ArduinoColab/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["arduino_colab_kernel", "arduino_colab_kernel.*"]

[tool.setuptools]
include-package-data = true